requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
PyMuPDF==1.24.0
Pillow==10.4.0
python-multipart==0.0.9
numpy>=1.24.0,<2.0
